reorganized sidebar files
