+++ app/src/main/java/com/bignerdranch/express/sidebarUi/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.bignerdranch.express.sidebarUi.mainline
+package com.bignerdranch.express.sidebarUi.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
