+++ app/src/main/java/com/bignerdranch/express/sidebarUi/profile/ProfileViewModel.kt
@@ -1,4 +1,4 @@
-package com.bignerdranch.express.sidebarUi.mainline
+package com.bignerdranch.express.sidebarUi.profile
 
 import androidx.lifecycle.ViewModel
