+++ app/src/main/res/navigation/sidebar_navigation.xml
@@ -13,7 +13,7 @@
 
     <fragment
         android:id="@+id/nav_profile"
-        android:name="com.bignerdranch.express.sidebarUi.mainline.ProfileFragment"
+        android:name="com.bignerdranch.express.sidebarUi.profile.ProfileFragment"
         android:label="@string/menu_profile"
         tools:layout="@layout/fragment_profile" />
